11.4.3
Refine `PermissionLevel` enum with new levels and descriptions.
Change `TotalPackets` type to `Int64` in `IDispatchChannel.cs`
and `DispatchRouter.cs` for larger packet counts. Update
`DispatchOptions.cs` to adjust `MaxPerConnectionQueue` validation
and add `BlockTimeout`. Refactor `DispatchChannel.cs` to improve
queue management, removing `TotalPackets` field and adding
`WaitForQueueSpace` method. Simplify connection management by
removing `OnConnectionClosed`. Improve worker management in
`PacketDispatchChannel.cs` with new fields and updated methods.

+++ src/Nalix.Network/Abstractions/IDispatchChannel.cs
@@ -19,7 +19,7 @@ public interface IDispatchChannel<TPacket> where TPacket : IPacket
     /// <value>
     /// The total number of packets currently enqueued for processing.
     /// </value>
-    System.Int32 TotalPackets { get; }
+    System.Int64 TotalPackets { get; }
 
     /// <summary>
     /// Adds a packet to the dispatch queue, associating it with a specific connection.
